# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bulid your own Supaglue",
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/customers": {
      "get": {
        "operationId": "mgmt-listCustomers",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/customer"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "mgmt-getCustomer",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/customer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.BAD_REQUEST"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.NOT_FOUND"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customer_id}": {
      "put": {
        "operationId": "mgmt-upsertCustomer",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": ["string", "null"]
                  },
                  "email": {
                    "type": ["string", "null"],
                    "description": "Email, but not validated as data from Supaglue has invalid emails for now"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/customer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.BAD_REQUEST"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.NOT_FOUND"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
    "/sync_configs": {
      "get": {
        "operationId": "mgmt-listSyncConfigs",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/sync_config"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
    "/verticals/sales-engagement/sequences": {
      "get": {
        "operationId": "salesEngagement-listSequences",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nextPageCursor": {
                      "type": ["string", "null"]
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/sales-engagement.sequence"
                      }
                    }
                  },
                  "required": ["items"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.BAD_REQUEST"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.NOT_FOUND"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
    "/verticals/sales-engagement/accounts/_upsert": {
      "post": {
        "operationId": "salesEngagement-upsertAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "record": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": ["string", "null"],
                        "example": "My Company"
                      },
                      "domain": {
                        "type": ["string", "null"],
                        "example": "mycompany.com"
                      },
                      "owner_id": {
                        "type": ["string", "null"],
                        "example": "9f3e97fd-4d5d-4efc-959d-bbebfac079f5"
                      },
                      "account_id": {
                        "type": ["string", "null"],
                        "example": "ae4be028-9078-4850-a0bf-d2112b7c4d11"
                      },
                      "custom_fields": {
                        "type": ["object", "null"],
                        "additionalProperties": {}
                      }
                    }
                  },
                  "upsert_on": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "The name of the account to upsert on. Supported for Outreach, Salesloft, and Apollo."
                      },
                      "domain": {
                        "type": "string",
                        "description": "The domain of the account to upsert on. Only supported for Outreach and Salesloft."
                      }
                    }
                  }
                },
                "required": ["record", "upsert_on"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "record": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      },
                      "required": ["id"]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.BAD_REQUEST"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
    "/verticals/sales-engagement/sequenceState": {
      "post": {
        "operationId": "salesEngagement-insertSequenceState",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "record": {
                    "type": "object",
                    "properties": {
                      "contact_id": {
                        "type": "string",
                        "example": "9f3e97fd-4d5d-4efc-959d-bbebfac079f5"
                      },
                      "mailbox_id": {
                        "type": "string",
                        "example": "ae4be028-9078-4850-a0bf-d2112b7c4d11"
                      },
                      "sequence_id": {
                        "type": "string",
                        "example": "b854e510-1c40-4ef6-ade4-8eb35f49d331"
                      },
                      "user_id": {
                        "type": ["string", "null"],
                        "example": "c590dc63-8e43-48a4-8154-1fbb00ac936b"
                      }
                    },
                    "required": ["contact_id", "mailbox_id", "sequence_id"]
                  }
                },
                "required": ["record"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "record": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      },
                      "required": ["id"]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.BAD_REQUEST"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
    "/verticals/crm/metadata/objects/{object_name}/properties": {
      "get": {
        "operationId": "crm-metadataListObjectProperties",
        "parameters": [
          {
            "in": "path",
            "name": "object_name",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/crm.meta.property"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.BAD_REQUEST"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.NOT_FOUND"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "x-customer-id": {
        "name": "x-customer-id",
        "type": "apiKey",
        "in": "header"
      },
      "x-provider-name": {
        "name": "x-provider-name",
        "type": "apiKey",
        "in": "header"
      },
      "x-nango-secret-key": {
        "name": "x-nango-secret-key",
        "type": "apiKey",
        "in": "header"
      },
      "x-api-key": {
        "name": "x-api-key",
        "type": "apiKey",
        "in": "header"
      },
      "x-use-new-backend": {
        "name": "x-use-new-backend",
        "type": "apiKey",
        "in": "header"
      }
    },
    "schemas": {
...
      "error.INTERNAL_SERVER_ERROR": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The error message",
            "example": "Internal server error"
          },
          "code": {
            "type": "string",
            "description": "The error code",
            "example": "INTERNAL_SERVER_ERROR"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "required": ["message"]
            },
            "description": "An array of issues that were responsible for the error",
            "example": []
          }
        },
        "required": ["message", "code"],
        "title": "Error",
        "description": "The error information",
        "example": {
          "code": "INTERNAL_SERVER_ERROR",
          "message": "Internal server error",
          "issues": []
        }
      },
      "customer": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string"
          },
          "name": {
            "type": ["string", "null"]
          },
          "email": {
            "type": ["string", "null"],
            "description": "Email, but not validated as data from Supaglue has invalid emails for now"
          }
        },
        "required": ["customer_id"]
      },
      "error.BAD_REQUEST": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The error message",
            "example": "Invalid input data"
          },
          "code": {
            "type": "string",
            "description": "The error code",
            "example": "BAD_REQUEST"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "required": ["message"]
            },
            "description": "An array of issues that were responsible for the error",
            "example": []
          }
        },
        "required": ["message", "code"],
        "title": "Error",
        "description": "The error information",
        "example": {
          "code": "BAD_REQUEST",
          "message": "Invalid input data",
          "issues": []
        }
      },
      "error.NOT_FOUND": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The error message",
            "example": "Not found"
          },
          "code": {
            "type": "string",
            "description": "The error code",
            "example": "NOT_FOUND"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "required": ["message"]
            },
            "description": "An array of issues that were responsible for the error",
            "example": []
          }
        },
        "required": ["message", "code"],
        "title": "Error",
        "description": "The error information",
        "example": {
          "code": "NOT_FOUND",
          "message": "Not found",
          "issues": []
        }
      },
...
      "sync_config": {
        "allOf": [
          {
            "$ref": "#/components/schemas/connection_sync_config"
          },
          {
            "type": "object",
            "properties": {
              "provider_name": {
                "type": "string"
              },
              "auto_start_on_connection": {
                "type": ["boolean", "null"],
                "description": "If true, the sync will start automatically when the connection is created."
              }
            },
            "required": ["provider_name"]
          }
        ]
      },
      "connection_sync_config": {
        "type": "object",
        "properties": {
          "destination_config": {
            "type": ["object", "null"],
            "properties": {
              "type": {
                "type": "string"
              },
              "schema": {
                "type": ["string", "null"]
              }
            },
            "required": ["type"]
          },
          "common_objects": {
            "type": ["array", "null"],
            "items": {
              "type": "object",
              "properties": {
                "object": {
                  "type": "string"
                }
              },
              "required": ["object"]
            }
          },
          "standard_objects": {
            "type": ["array", "null"],
            "items": {
              "type": "object",
              "properties": {
                "object": {
                  "type": "string"
                }
              },
              "required": ["object"]
            }
          },
          "custom_objects": {
            "type": ["array", "null"],
            "items": {
              "type": "object",
              "properties": {
                "object": {
                  "type": "string"
                }
              },
              "required": ["object"]
            }
          }
        }
      },
...
      "sales-engagement.sequence": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "is_deleted": {
            "type": "boolean"
          },
          "last_modified_at": {
            "type": "string"
          },
          "owner_id": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "num_steps": {
            "type": "number"
          },
          "metrics": {
            "type": "object",
            "additionalProperties": {}
          },
          "is_enabled": {
            "type": "boolean"
          },
          "raw_data": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "is_deleted",
          "last_modified_at",
          "owner_id",
          "tags",
          "num_steps",
          "is_enabled"
        ]
      },
...
      "crm.meta.property": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The machine name of the property as it appears in the third-party Provider",
            "example": "FirstName"
          },
          "label": {
            "type": "string",
            "description": "The human-readable name of the property as provided by the third-party Provider.",
            "example": "First Name"
          },
          "type": {
            "type": "string",
            "description": "The type of the property as provided by the third-party Provider. These types are not unified by Supaglue. For Intercom, this is string, integer, boolean, or object. For Outreach, this is integer, boolean, number, array, or string.",
            "example": "string"
          },
          "raw_details": {
            "type": "object",
            "additionalProperties": {},
            "description": "The raw details of the property as provided by the third-party Provider, if available.",
            "example": {}
          }
        },
        "required": ["id", "label"]
      },
...
    }
  }
}